pyproject.toml
[project]
name = "uralsteel-engineer-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "lightrag-api>=0.1.4",
    "psutil>=7.2.2",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
